sage/requirements.txt
Django>=5.2
psycopg2-binary
python-dotenv
djangorestframework>=3.14
django-filter>=23.0